Moved implementation to cpp file
